fix: widen setKeyId id parameter to size_t so narrow-before-check is impossible
PR review feedback on Phase A: the previous fix kept id as uint8_t, so
callers passing wider types (size_t loop indices, int config.id) had
their values silently narrowed at the call site before the id > 99
guard could run. Values of 256 would wrap to 0 and masquerade as peer 0.

Widening to size_t means the check runs on the full caller-supplied
value; only then are the two digits computed. The format specifier
updates from %u to %zu to match.

+++ lib/AstrOsStorageManager/src/NvsManager.c
@@ -9,11 +9,14 @@
 
 static const char *TAG = "NvsManager";
 
-static bool setKeyId(char *key, uint8_t id, uint8_t startPos)
+// id is size_t (not uint8_t) so callers passing loop indices don't silently
+// narrow before the range check: a value of 256 must fail the guard, not
+// wrap to 0 and masquerade as a valid peer 0 index.
+static bool setKeyId(char *key, size_t id, uint8_t startPos)
 {
     if (id > 99)
     {
-        ESP_LOGE(TAG, "Peer index %u exceeds maximum of 99", (unsigned)id);
+        ESP_LOGE(TAG, "Peer index %zu exceeds maximum of 99", id);
         return false;
     }
     key[startPos] = '0' + (id / 10);
